feat(common/extension/boolean): add package docs and deprecate BoolToStr
- Introduce doc.go with package-level overview
- Add README.md with features, quick start, and guides
- Provide example_test.go with runnable GoDoc examples
- Move BoolToStr helper from common to extension/boolean and rename to BoolToString
- Mark BoolToStr as deprecated in favor of BoolToString
